Update with new URL for overview page
I opted to log in manually to mint, and after doing so successfully, `mint-amazon-tagger` did not recognize that I was logged in.

It looks like the test to determine if a user is logged into mint is the URL of the overview page. The URL I guess changed in a recent redesign from `overview.event` to `overview`. Unsure if other changes are required.